Added full functionality

+++ src/main/java/com/leverx/dealerstat/entity/Comment.java
@@ -6,7 +6,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
-import java.util.Date;
+import java.sql.Date;
 
 @Entity
 @Table(name = "comment", catalog = "dealerstatdb")
@@ -17,7 +17,7 @@ public class Comment {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String text;
-    private int rating;
+    private double rating;
     @CreationTimestamp
     @Column(name = "created_at")
     private Date dateOfCreation;
@@ -46,11 +46,11 @@ public class Comment {
         this.text = text;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
